components share state to toggle active tab
