+++ src/css/custom.css
@@ -591,6 +591,10 @@ a:hover {
     color: black;
     background-color: #f1c00f;
 }
+.home-tab {
+    position: absolute;
+    left: 100px;
+}
 
 /***************/
 /* Map Overlay */
